requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
Pillow>=8.2.0
tqdm>=4.61.0
matplotlib>=3.4.2
scikit-image>=0.18.1
opencv-python>=4.5.2
lpips>=0.1.4
tensorboard>=2.6.0 